services/validation-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
httpx==0.28.1
pydantic==2.10.4
PyJWT==2.10.1
prometheus-client==0.21.1
